refactor: Split up interest calculator into smaller components
